# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tornería Montero API",
    "version": "1.0.0",
    "description": "API para gestionar trabajos, clientes y trabajadores de Tornería Montero. Autenticación por header x-api-key."
  },
  "servers": [
    { "url": "/", "description": "Local server" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Legacy API key header (x-api-key). BearerAuth (JWT) is preferred."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "UnifiedError": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Mensaje de error legible" }
        }
      },
...
      "ClientListResponseWrapper": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Client" } },
          "error": { "$ref": "#/components/schemas/UnifiedError", "nullable": true }
        }
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "ci_rut": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "telefono": { "type": "string" },
          "direccion": { "type": "string" },
          "estado": { "type": "string" },
          "fecha_registro": { "type": "string", "format": "date-time" }
        }
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "ci": { "type": "string" },
          "rol_tecnico": { "type": "string" },
          "estado": { "type": "string" },
          "usuario_id": { "type": "integer" }
        }
      },
      "PedidoCreate": {
        "type": "object",
        "required": ["descripcion", "cliente_id"],
        "properties": {
          "descripcion": { 
            "type": "string",
            "description": "Descripción detallada del pedido"
          },
          "prioridad": { 
            "type": "string", 
            "enum": ["BAJA", "MEDIA", "ALTA"], 
            "default": "MEDIA",
            "description": "Prioridad del pedido"
          },
          "cliente_id": { 
            "type": "integer",
            "description": "ID del cliente que realiza el pedido"
          },
          "responsable_id": { 
            "type": ["integer", "null"],
            "description": "ID del trabajador asignado al pedido (opcional)"
          },
          "fecha_estimada_fin": { 
            "type": "string", 
            "format": "date",
            "description": "Fecha estimada de finalización (YYYY-MM-DD)"
          },
          "precio": { 
            "type": ["number", "string"], 
            "description": "Precio del trabajo"
          },
          "notas": {
            "type": "string",
            "description": "Notas adicionales sobre el trabajo (opcional)"
          },
          "tiempo_estimado_sec": {
            "type": "integer",
            "description": "Tiempo estimado en segundos (opcional)"
          },
          "adjuntos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de URLs o identificadores de archivos adjuntos (opcional)"
          }
        }
      },
      "PedidoResponse": {
        "type": "object",
        "properties": {
          "id": { 
            "type": "integer",
            "description": "ID único del pedido" 
          },
          "cliente_id": {
            "type": "integer",
            "description": "ID del cliente que realizó el pedido"
          },
          "descripcion": { 
            "type": "string",
            "description": "Descripción detallada del pedido"
          },
          "prioridad": { 
            "type": "string", 
            "enum": ["BAJA", "MEDIA", "ALTA"],
            "description": "Nivel de prioridad del pedido"
          },
          "precio": { 
            "type": "string",
            "description": "Precio del trabajo"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de inicio del trabajo"
          },
          "fecha_estimada_fin": { 
            "type": "string", 
            "format": "date-time",
            "description": "Fecha y hora estimada de finalización"
          },
          "estado": { 
            "type": "string", 
            "enum": ["PENDIENTE", "ASIGNADO", "EN_PROGRESO", "QA", "ENTREGADO"],
            "description": "Estado actual del pedido"
          },
          "responsable_id": { 
            "type": ["integer", "null"],
            "description": "ID del trabajador asignado (si existe)"
          },
          "semaforo": {
            "type": "string",
            "enum": ["VERDE", "AMARILLO", "ROJO"],
            "description": "Estado del semáforo para seguimiento"
          },
          "notas": {
            "type": ["string", "null"],
            "description": "Notas adicionales sobre el trabajo"
          },
          "adjuntos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de archivos adjuntos"
          },
          "tiempo_estimado_sec": {
            "type": ["integer", "null"],
            "description": "Tiempo estimado en segundos"
          },
          "tiempo_real_sec": {
            "type": ["integer", "null"],
            "description": "Tiempo real empleado en segundos"
          },
          "creado_por_id": {
            "type": ["integer", "null"],
            "description": "ID del usuario que creó el pedido"
          },
          "actualizado_por_id": {
            "type": ["integer", "null"],
            "description": "ID del último usuario que actualizó el pedido"
          },
          "fecha_actualizacion": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de la última actualización"
          },
          "eliminado": {
            "type": "boolean",
            "description": "Indica si el pedido ha sido marcado como eliminado"
          }
        },
        "required": ["id", "cliente_id", "descripcion", "prioridad", "estado", "fecha_actualizacion"]
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
    "/api/jobs": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar trabajos (compatibilidad)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de trabajos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PedidoResponse" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Pedidos"],
        "summary": "Crear trabajo (compatibilidad)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PedidoCreate" } } }
        },
        "responses": {
          "201": { "description": "Creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PedidoResponse" } } } }
        }
      }
    },
    "/api/jobs/workers": {
      "get": {
        "tags": ["Trabajadores"],
        "summary": "Listar trabajadores (desde jobs)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Worker" } } } } }
        }
      }
    },
...
    "/api/trabajadores": {
      "get": {
        "tags": ["Trabajadores"],
        "summary": "Listar trabajadores activos",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Worker" } } }
            }
          }
        }
      },
      "post": {
        "tags": ["Trabajadores"],
        "summary": "Crear un trabajador",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ci","usuario_id"],
                "properties": {
                  "ci": { "type": "string" },
                  "rol_tecnico": { "type": "string" },
                  "usuario_id": { "type": "integer" },
                  "direccion": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Trabajador creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Worker" } } } }
        }
      }
    },
    "/api/trabajadores/{id}": {
      "get": {
        "tags": ["Trabajadores"],
        "summary": "Obtener trabajador por ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Worker" } } } }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Trabajadores"],
        "summary": "Actualizar trabajador",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "rol_tecnico": { "type": "string" }, "estado": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Worker" } } } }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Trabajadores"],
        "summary": "Eliminar trabajador",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Eliminado", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } } }, "404": { "description": "Not found" } }
      }
    },
...
    "/api/clientes": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Listar todos los clientes",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { 
            "name": "query", 
            "in": "query", 
            "required": false, 
            "schema": { "type": "string" },
            "description": "Término de búsqueda opcional para filtrar clientes"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClientListResponseWrapper" }
              }
            }
          },
          "401": { 
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Clientes"],
        "summary": "Crear un nuevo cliente",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "ci_rut"],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre completo del cliente"
                  },
                  "ci_rut": {
                    "type": "string",
                    "description": "Cédula o RUT del cliente"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico del cliente"
                  },
                  "telefono": {
                    "type": "string",
                    "description": "Número de teléfono del cliente"
                  },
                  "direccion": {
                    "type": "string",
                    "description": "Dirección del cliente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente creado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Client" }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Obtener un cliente por ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID del cliente a obtener"
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Client" }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Clientes"],
        "summary": "Actualizar un cliente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID del cliente a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre completo del cliente"
                  },
                  "ci_rut": {
                    "type": "string",
                    "description": "Cédula o RUT del cliente"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico del cliente"
                  },
                  "telefono": {
                    "type": "string",
                    "description": "Número de teléfono del cliente"
                  },
                  "direccion": {
                    "type": "string",
                    "description": "Dirección del cliente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Client" }
              }
            }
          },
          "401": {
            "description": "No autenticado o no autorizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Clientes"],
        "summary": "Eliminar un cliente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID del cliente a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cliente eliminado correctamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado o no autorizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}